Fix 1.2 - Фикс detect
