+++ app/src/main/java/cy/volleybolley/core/presentation/ui/screens/authorization/OnboardingScreen.kt
@@ -10,12 +10,9 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavHostController
-import androidx.navigation.compose.rememberNavController
 import cy.volleybolley.core.presentation.ui.navigation.NavMap
-import cy.volleybolley.ui.theme.VolleybolleyTheme
 
 @Composable
 fun OnboardingScreen(navController: NavHostController) {
