--- app/src/main/java/cy/volleybolley/core/presentation/ui/screens/createNewGame/BasicGameSetupScreen.kt
@@ -1,22 +1,0 @@
-package cy.volleybolley.core.presentation.ui.screens.createNewGame
-
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Button
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
-import androidx.navigation.NavHostController
-import cy.volleybolley.core.presentation.ui.navigation.NavMap
-
-@Composable
-fun BasicGameSetupScreen(navController: NavHostController) {
-    Button(onClick = { navController.popBackStack() }) {
-        Text("Назад")
-    }
-}
