--- app/src/main/java/cy/volleybolley/core/presentation/ui/screens/findAGame/JoinTheGameScreen.kt
@@ -1,13 +1,0 @@
-package cy.volleybolley.core.presentation.ui.screens.findAGame
-
-import androidx.compose.material3.Button
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.navigation.NavHostController
-
-@Composable
-fun JoinTheGameScreen(navController: NavHostController) {
-    Button(onClick = { navController.popBackStack() }) {
-        Text("Назад")
-    }
-}
