--- app/src/main/java/cy/volleybolley/core/presentation/ui/screens/games/gameInvites/GameInvitesScreen.kt
@@ -1,13 +1,0 @@
-package cy.volleybolley.core.presentation.ui.screens.games.gameInvites
-
-import androidx.compose.material3.Button
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.navigation.NavHostController
-
-@Composable
-fun GameInvitesScreen(navController: NavHostController) {
-    Button(onClick = { navController.popBackStack() }) {
-        Text("Назад")
-    }
-}
